clean up logging
